Print line if we have not seen any pids created yet
This specifically handles the case where you get a log snippet that may
not have any pid creation lines in it and pidcat ends up filtering
everything out. In this case we just print the line if we haven't seen
any pids yet.

An alternative would be allowing filter by pid, however I find grep
works fine for that.