pidcat: print a newline when SIGINT is sent
Ctrl + C/Cmd + C is, what I believe, a pretty common way of exiting from pidcat, but
the default behavior results in prompts getting messed up because there is no final newline.
ZSH seems to work around it by indicating that there's a missing newline in the output
and adding one itself, but most other shell/prompt combinations do not have this, so
we need this as a workaround.

Signed-off-by: Harsh Shandilya <me@msfjarvis.dev>